requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.16.1
fsspec==2024.9.0
idna==3.10
jinja2==3.1.4
keyboard==0.13.5
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
PyAudio==0.2.14
regex==2024.9.11
requests==2.32.3
sympy==1.13.3
tiktoken==0.8.0
torch==2.4.1
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.3
